# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FaceTrack API
  description: 'FaceTrack attendance system API (requirements §12). Portal: session
    cookie + CSRF header; integrations: X-API-Key.'
  version: 1.0.0
paths:
...
  /api/v1/employees:
    get:
      tags:
      - employees
      summary: List employees
      operationId: list_employees
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Search
      - name: department_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Department Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/EmployeeStatus'
          - type: 'null'
          title: Status
      - name: enrolled
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enrolled
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
            pattern: ^-?[a-z_]+$
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_EmployeeOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - employees
      summary: Create Employee
      operationId: create_employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_EmployeeOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{employee_id}:
    get:
      tags:
      - employees
      summary: Get an employee
      operationId: get_employee
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_EmployeeOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - employees
      summary: Update Employee
      description: Setting status to inactive stops recognition immediately (FR-13).
      operationId: update_employee
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_EmployeeOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - employees
      summary: Delete an employee and permanently erase their face data
      description: 'FR-13: the record is soft-deleted (attendance history stays) and
        all biometric data is hard-deleted.

        `confirm_employee_code` must equal the employee''s code.'
      operationId: delete_employee
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      - name: confirm_employee_code
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          title: Confirm Employee Code
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/employees/{employee_id}/faces:
    get:
      tags:
      - employees
      summary: Face gallery
      operationId: list_faces
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_list_FaceOut__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - employees
      summary: Upload enrollment photos (multipart)
      description: 'FR-8/FR-9: each photo is checked (one face, width >= 112 px, quality,
        blur) and either stored

        or rejected with a reason. Requires recorded consent (422 otherwise); 503
        if the engine is down.

        FR-10: `possible_duplicates` lists other employees with a very similar face.'
      operationId: upload_faces
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_faces'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_EnrollmentResult_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{employee_id}/faces/{face_id}:
    delete:
      tags:
      - employees
      summary: Delete Face
      operationId: delete_face
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      - name: face_id
        in: path
        required: true
        schema:
          type: integer
          title: Face Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{employee_id}/faces/{face_id}/image:
    get:
      tags:
      - employees
      summary: Enrollment photo
      operationId: face_image
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      - name: face_id
        in: path
        required: true
        schema:
          type: integer
          title: Face Id
      - name: thumbnail
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Thumbnail
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{employee_id}/photo:
    get:
      tags:
      - employees
      summary: Best enrollment photo (thumbnail)
      operationId: employee_photo
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/import:
    post:
      tags:
      - employees
      summary: Bulk import employees (Excel/CSV) and photos (ZIP named by employee
        code)
      description: Runs in the background (202). Poll `GET /jobs/{job_id}` for the
        per-row result.
      operationId: import_employees
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_employees'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ImportAccepted_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/leaves:
    get:
      tags:
      - employees
      summary: List leave
      operationId: list_leaves
      parameters:
      - name: employee_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Employee Id
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date To
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
            pattern: ^-?[a-z_]+$
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_LeaveOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - employees
      summary: Create Leave
      operationId: create_leave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_LeaveOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/leaves/{leave_id}:
    delete:
      tags:
      - employees
      summary: Delete Leave
      operationId: delete_leave
      parameters:
      - name: leave_id
        in: path
        required: true
        schema:
          type: integer
          title: Leave Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/jobs/{job_id}:
    get:
      tags:
      - system
      summary: Background job status
      operationId: get_job
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_JobOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiResponse_EmployeeOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/EmployeeOut'
      type: object
      required:
      - message
      - data
      title: ApiResponse[EmployeeOut]
    ApiResponse_EnrollmentResult_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/EnrollmentResult'
      type: object
      required:
      - message
      - data
      title: ApiResponse[EnrollmentResult]
...
    ApiResponse_ImportAccepted_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/ImportAccepted'
      type: object
      required:
      - message
      - data
      title: ApiResponse[ImportAccepted]
    ApiResponse_JobOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/JobOut'
      type: object
      required:
      - message
      - data
      title: ApiResponse[JobOut]
    ApiResponse_LeaveOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/LeaveOut'
      type: object
      required:
      - message
      - data
      title: ApiResponse[LeaveOut]
...
    ApiResponse_list_FaceOut__:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          items:
            $ref: '#/components/schemas/FaceOut'
          type: array
          title: Data
      type: object
      required:
      - message
      - data
      title: ApiResponse[list[FaceOut]]
...
    Body_import_employees:
      properties:
        sheet:
          type: string
          contentMediaType: application/octet-stream
          title: Sheet
          description: XLSX or CSV, <= 10 MB
        photos_zip:
          anyOf:
          - type: string
            contentMediaType: application/octet-stream
          - type: 'null'
          title: Photos Zip
          description: ZIP of <employee_code>/<n>.jpg, <= 200 MB
      type: object
      required:
      - sheet
      title: Body_import_employees
    Body_upload_faces:
      properties:
        photos:
          items:
            type: string
            contentMediaType: application/octet-stream
          type: array
          title: Photos
          description: JPEG/PNG, <= 5 MB each, 1-10 files
        source:
          type: string
          pattern: ^(upload|webcam)$
          title: Source
          default: upload
      type: object
      required:
      - photos
      title: Body_upload_faces
...
    EmployeeCreate:
      properties:
        employee_code:
          type: string
          maxLength: 64
          minLength: 1
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$
          title: Employee Code
        full_name:
          type: string
          maxLength: 200
          minLength: 1
          title: Full Name
        department_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Department Id
        designation:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Designation
        shift_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Shift Id
        location_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Location Id
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Phone
        status:
          $ref: '#/components/schemas/EmployeeStatus'
          default: active
        consent_signed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Consent Signed At
        hr_external_id:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Hr External Id
      additionalProperties: false
      type: object
      required:
      - employee_code
      - full_name
      title: EmployeeCreate
    EmployeeOut:
      properties:
        id:
          type: integer
          title: Id
        employee_code:
          type: string
          title: Employee Code
        full_name:
          type: string
          title: Full Name
        department_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Department Id
        department_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Department Name
        designation:
          anyOf:
          - type: string
          - type: 'null'
          title: Designation
        shift_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Shift Id
        shift_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Shift Name
        location_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Location Id
        location_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Name
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        status:
          $ref: '#/components/schemas/EmployeeStatus'
        consent_signed_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Consent Signed At
        deactivated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Deactivated At
        hr_external_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Hr External Id
        enrolled_photos:
          type: integer
          title: Enrolled Photos
          default: 0
        enrollment_complete:
          type: boolean
          title: Enrollment Complete
          default: false
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - employee_code
      - full_name
      - department_id
      - designation
      - shift_id
      - location_id
      - email
      - phone
      - status
      - consent_signed_at
      - deactivated_at
      - hr_external_id
      - created_at
      title: EmployeeOut
    EmployeeStatus:
      type: string
      enum:
      - active
      - inactive
      title: EmployeeStatus
    EmployeeUpdate:
      properties:
        full_name:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Full Name
        department_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Department Id
        designation:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Designation
        shift_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Shift Id
        location_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Location Id
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Phone
        status:
          anyOf:
          - $ref: '#/components/schemas/EmployeeStatus'
          - type: 'null'
        consent_signed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Consent Signed At
        hr_external_id:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Hr External Id
      additionalProperties: false
      type: object
      title: EmployeeUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LeaveCreate:
      properties:
        employee_id:
          type: integer
          title: Employee Id
        from_date:
          type: string
          format: date
          title: From Date
        to_date:
          type: string
          format: date
          title: To Date
        type:
          type: string
          maxLength: 64
          minLength: 1
          title: Type
      additionalProperties: false
      type: object
      required:
      - employee_id
      - from_date
      - to_date
      - type
      title: LeaveCreate
...
    PaginatedResponse_EmployeeOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          items:
            $ref: '#/components/schemas/EmployeeOut'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - message
      - data
      - pagination
      title: PaginatedResponse[EmployeeOut]
...
    PaginatedResponse_LeaveOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          items:
            $ref: '#/components/schemas/LeaveOut'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - message
      - data
      - pagination
      title: PaginatedResponse[LeaveOut]
...
    Pagination:
      properties:
        page:
          type: integer
          minimum: 1.0
          title: Page
        page_size:
          type: integer
          minimum: 1.0
          title: Page Size
        total:
          type: integer
          minimum: 0.0
          title: Total
      type: object
      required:
      - page
      - page_size
      - total
      title: Pagination
...